Add allowed hosts to web dev server
This is a list of Reserved TLDs that will never be valid TLDs in a
global DNS sense [1].

Previously, we suggested to use `dev` which is now a valid global TLD
with mandatory redirect to HTTPS in latest version of Chrome and FF.
Which makes it complicated to use for local development.

The documentation will be updated in the following commit as well.

[1]: https://iyware.com/dont-use-dev-for-development/